refactor(routes): extract playback control routes
