+++ app/relaytv_app/routes/__init__.py
@@ -31,6 +31,21 @@ from .app_info import router as app_info_router
 from .assets import _resolve_static_asset, router as assets_router
 from .devices import router as devices_router
 from .health import router as health_router
+from .playback import (
+    MuteReq as MuteReq,
+    SeekAbsReq as SeekAbsReq,
+    SeekReq as SeekReq,
+    VolumeReq as VolumeReq,
+    mute as mute,
+    pause as pause,
+    playback_state as playback_state,
+    resume as resume,
+    seek as seek,
+    seek_abs as seek_abs,
+    toggle_pause as toggle_pause,
+    volume as volume,
+)
+from .playback import router as playback_router
 from .queue import router as queue_router
 from .snapshots import router as snapshots_router
 from .status import router as status_router
@@ -42,6 +57,7 @@ router.include_router(app_info_router)
 router.include_router(assets_router)
 router.include_router(devices_router)
 router.include_router(health_router)
+router.include_router(playback_router)
 router.include_router(queue_router)
 router.include_router(snapshots_router)
 router.include_router(status_router)
@@ -109,23 +125,6 @@ class PlayReq(BaseModel):
     url: str
     use_ytdlp: bool = True
     cec: bool = False  # default false (most setups don't have /dev/cec0)
-
-
-class VolumeReq(BaseModel):
-    delta: float | None = None
-    set: float | None = None
-
-
-class MuteReq(BaseModel):
-    set: bool | None = None
-
-
-class SeekReq(BaseModel):
-    sec: float
-
-
-class SeekAbsReq(BaseModel):
-    sec: float
 
 
 class PlayNowReq(BaseModel):
@@ -6789,32 +6788,6 @@ def _seek_absolute_result(target_sec: float) -> dict[str, object]:
     if isinstance(result, dict):
         return result
     return _control_result_or_raise(player.mpv_command(["seek", float(target_sec), "absolute"]), action="seek_abs")
-@router.post("/pause")
-def pause():
-    result = _control_result_or_raise(player.mpv_set_result("pause", True), action="pause")
-    state.set_session_state("paused")
-    state.set_pause_reason("user")
-    return {"ok": True, "paused": True, **_control_ack_payload(result)}
-
-
-@router.post("/resume")
-def resume():
-    result = _control_result_or_raise(player.mpv_set_result("pause", False), action="resume")
-    state.set_session_state("playing")
-    state.set_pause_reason(None)
-    return {"ok": True, "paused": False, **_control_ack_payload(result)}
-
-
-@router.post("/toggle_pause")
-def toggle_pause():
-    cur = bool(player.mpv_get("pause"))
-    target = not cur
-    result = _control_result_or_raise(player.mpv_set_result("pause", target), action="toggle_pause")
-    state.set_session_state("paused" if target else "playing")
-    state.set_pause_reason("user" if target else None)
-    return {"ok": True, "paused": target, **_control_ack_payload(result)}
-
-
 @router.post("/playback/play")
 def playback_play():
     """
@@ -6918,62 +6891,6 @@ def playback_toggle():
         state.set_pause_reason("user" if target else None)
         return {"ok": True, "action": "toggle_pause", "paused": target, **_control_ack_payload(result)}
     return playback_play()
-
-
-
-@router.post("/seek")
-def seek(req: SeekReq):
-    hold_sec = _seek_transition_hold_sec()
-    state.AUTO_NEXT_SUPPRESS_UNTIL = max(float(getattr(state, "AUTO_NEXT_SUPPRESS_UNTIL", 0.0) or 0.0), time.time() + hold_sec)
-    try:
-        player._mark_playback_transition(hold_sec)
-    except Exception:
-        pass
-    result = _seek_relative_result(float(req.sec))
-    return {"ok": True, "seeked": req.sec, **_control_ack_payload(result)}
-
-
-@router.post("/seek_abs")
-def seek_abs(req: SeekAbsReq):
-    # Seek to absolute position (seconds)
-    hold_sec = _seek_transition_hold_sec()
-    state.AUTO_NEXT_SUPPRESS_UNTIL = max(float(getattr(state, "AUTO_NEXT_SUPPRESS_UNTIL", 0.0) or 0.0), time.time() + hold_sec)
-    try:
-        player._mark_playback_transition(hold_sec)
-    except Exception:
-        pass
-    result = _seek_absolute_result(float(req.sec))
-    return {"ok": True, "seeked_to": req.sec, **_control_ack_payload(result)}
-
-@router.post("/volume")
-def volume(req: VolumeReq):
-    if req.set is not None:
-        v = max(0.0, min(200.0, float(req.set)))
-        result = _control_result_or_raise(player.mpv_set_result("volume", v), action="volume")
-        state.update_settings({"volume": v})
-        return {"ok": True, "volume": v, **_control_ack_payload(result)}
-    if req.delta is not None:
-        cur = float(player.mpv_get("volume") or 0.0)
-        v = max(0.0, min(200.0, cur + float(req.delta)))
-        result = _control_result_or_raise(player.mpv_set_result("volume", v), action="volume")
-        state.update_settings({"volume": v})
-        return {"ok": True, "volume": v, **_control_ack_payload(result)}
-    raise HTTPException(status_code=400, detail="Provide delta or set")
-
-
-@router.post("/mute")
-def mute(req: MuteReq):
-    """Toggle mute (or explicitly set it) using mpv's native mute property."""
-    try:
-        cur = bool(player.mpv_get("mute"))
-    except Exception:
-        cur = False
-    target = (not cur) if req.set is None else bool(req.set)
-    try:
-        result = _control_result_or_raise(player.mpv_set_result("mute", target), action="mute")
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"mute failed: {e}")
-    return {"ok": True, "mute": target, **_control_ack_payload(result)}
 
 
 
@@ -7455,12 +7372,6 @@ def _playback_state_fast_snapshot() -> dict[str, object]:
     )
     payload.update(state.update_playback_runtime_state(runtime_state, runtime_reason))
     return payload
-
-
-@router.get("/playback/state")
-def playback_state():
-    """Lightweight playback state for overlay/browser polling."""
-    return _playback_state_fast_snapshot()
 
 
 def _status_payload() -> dict[str, object]:
